==> code/00-SMA-data-gathering.ipynb <==
### Libraries, paths, and set-up:

import requests
import os
from selenium import webdriver
from selenium.webdriver.common.action_chains import ActionChains
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
import time
os.chdir('/Users/manotas/Documents/GitHub-Repos/ML-Energy-Colombia')

paths = ['prices/supplyprice','prices/stockprice/national_stock','prices/marginalcost',
         'availability','resource','capacity','generation']

for i in paths:
    if not os.path.exists('data/external/'+i):
        os.mkdir('data/external/'+i)

### How to download iterative .xlsx files from source

# Constructing a Selenium parser for files:

# A new instance of the Firefox driver
driver = webdriver.Firefox()

# Navigates to the webpage (this example is for)
driver.get('https://sinergox.xm.com.co/trpr/Paginas/Historicos/Historicos.aspx?RootFolder=%2Ftrpr%2FHistricos%2FContratos&FolderCTID=0x012000394993FA303733428C33EC91D1DFA6DB&View=%7B5CA2173E-1541-4ec7-9D1C-E145E3DFFAE3%7D#InplviewHash5ca2173e-1541-4ec7-9d1c-e145e3dffae3=Paged%3DTRUE-p_SortBehavior%3D0-p_FileLeafRef%3DVentas%255fContrato%255f%2528kwh%2529%255f2007%252exlsx-p_ID%3D144-PageFirstRow%3D31')

# Waits for the page to load
time.sleep(5)

# Finds all the links on the page
all_links = driver.find_elements(By.TAG_NAME, 'a')

# Filters the links based on the text
links = [link for link in all_links if 'Precio_Bolsa_Nacional_($kwh)' in link.text]

# Creates an instance of ActionChains
actions = ActionChains(driver)

# Iterates over the links
for link in links:
    # Right-clicks each link
    actions.context_click(link).perform()
    # Waits for the context menu to appear
    time.sleep(10)
    # Clicks the "Descargar" option
    descargar = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.LINK_TEXT, 'Descargar')))
    descargar.click()
    # Waits for the download to complete
    time.sleep(300)

# Closes the browser
driver.quit()
