==> code/09-SMA-ML-noverlapping.ipynb <==
### 0. Libraries, paths, and set-up

import pandas as pd
import numpy as np
import os
import pickle
import tensorflow as tf
import keras
import sys
from tensorflow.keras.layers import Layer, Dense, LSTM, GRU, Dropout, Input, Concatenate, Bidirectional, Attention
from tensorflow.keras.models import Model
from tensorflow.keras.optimizers.legacy import Adam
from tensorflow.keras.callbacks import EarlyStopping
from tensorflow.keras.models import load_model
import keras_tuner as kt
from tqdm import tqdm
from sklearn.preprocessing import StandardScaler
import keras_tuner as kt

# For reproducibility
tf.random.set_seed(42)
np.random.seed(42)

# Setting directory
os.chdir('/Users/manotas/Documents/GitHub-Repos/ML-Energy-Colombia')

# Custom Modules
from src.models.metrics import calculate_metrics

# Loading the windowed data
with open('data/processed/train_non_overlapping_windows.pkl', 'rb') as f:
    X_train_windows, Y_train_windows = pickle.load(f)
with open('data/processed/test_non_overlapping_windows.pkl', 'rb') as f:
    X_test_windows, Y_test_windows = pickle.load(f)

# Storing references for plant and agent names
train_references = [(window['plant'].iloc[0], window['agent'].iloc[0]) for window in X_train_windows]
test_references = [(window['plant'].iloc[0], window['agent'].iloc[0]) for window in X_test_windows]

# Dropping 'datetime', 'plant', and 'agent' columns and convert data types to float
for window in X_train_windows:
    window.drop(columns=['datetime', 'plant', 'agent'], errors='ignore', inplace=True)
    window.loc[:, :] = window.astype(float)  # Convert all remaining columns to float

for window in X_test_windows:
    window.drop(columns=['datetime', 'plant', 'agent'], errors='ignore', inplace=True)
    window.loc[:, :] = window.astype(float)  # Converting all remaining columns to float

# Converting to numpy arrays (as RNNs require 3D input)
X_train_np = np.array([window.values for window in X_train_windows])
Y_train_np = np.array([window for window in Y_train_windows])
X_test_np = np.array([window.values for window in X_test_windows])
Y_test_np = np.array([window for window in Y_test_windows])


# Checking data types and shapes
print("X_train_np dtype:", X_train_np.dtype)
print("Y_train_np dtype:", Y_train_np.dtype)
print("X_train_np shape:", X_train_np.shape)
print("Y_train_np shape:", Y_train_np.shape)

# Ensuring the data is of type float32
X_train_np = X_train_np.astype(np.float32)
Y_train_np = Y_train_np.astype(np.float32)
X_test_np = X_test_np.astype(np.float32)
Y_test_np = Y_test_np.astype(np.float32)

# Verifying the conversion
print("X_train_np dtype after conversion:", X_train_np.dtype)
print("Y_train_np dtype after conversion:", Y_train_np.dtype)


#### Ensuring GPU availability

print(f"Tensor Flow Version: {tf.__version__}")
print(f"Keras Version: {keras.__version__}")
print()
print(f"Python {sys.version}")
gpu = len(tf.config.list_physical_devices('GPU'))>0
print()
print("GPU is", "AVAILABLE" if gpu else "NOT AVAILABLE")

### 1. RNN model set-up

#### T2V Layer

class Time2Vec(Layer):
    def __init__(self, kernel_size=1, **kwargs):
        self.kernel_size = kernel_size
        super(Time2Vec, self).__init__(**kwargs)

    def build(self, input_shape):
        self.W = self.add_weight(name='W',
                                 shape=(input_shape[-1], self.kernel_size),
                                 initializer='uniform',
                                 trainable=True)
        self.B = self.add_weight(name='B',
                                 shape=(1, self.kernel_size),
                                 initializer='uniform',
                                 trainable=True)
        self.w = self.add_weight(name='w',
                                 shape=(input_shape[-1],),
                                 initializer='uniform',
                                 trainable=True)
        self.b = self.add_weight(name='b',
                                 shape=(input_shape[-1],),
                                 initializer='uniform',
                                 trainable=True)
        super(Time2Vec, self).build(input_shape)

    def call(self, inputs):
        bias = self.w * inputs + self.b
        dp = tf.matmul(inputs, self.W) + self.B
        wgts = tf.math.sin(dp)
        return tf.concat([wgts, bias], -1)

    def compute_output_shape(self, input_shape):
        return (input_shape[0], input_shape[1], 2 * input_shape[2])

#### Defining with hyperparameter options for tuning

def build_model(hp):
    inputs = Input(shape=(24, X_train_np.shape[2]))  # Assuming 24 hours and the preprocessed shape
    time2vec = Time2Vec(kernel_size=hp.Int('t2v_kernel_size', min_value=1, max_value=5))(inputs)
    
    x = time2vec
    
    for i in range(hp.Int('num_layers', 1, 3)):
        if hp.Choice('rnn_type', ['LSTM', 'GRU']) == 'LSTM':
            x = Bidirectional(LSTM(units=hp.Int('units', min_value=32, max_value=128, step=32), 
                                   return_sequences=True))(x)
        else:
            x = Bidirectional(GRU(units=hp.Int('units', min_value=32, max_value=128, step=32), 
                                  return_sequences=True))(x)
        if hp.Boolean('dropout'):
            x = Dropout(rate=0.2)(x)

    # Adding Attention Mechanism
    attention = Attention()([x, x])
    
    # Reduce to single output
    x = Dense(1)(attention)
    
    model = Model(inputs, x)
    model.compile(optimizer=Adam(learning_rate=hp.Choice('learning_rate', [1e-2, 1e-3, 1e-4])),
                  loss='mse', 
                  metrics=['mae'])
    return model

# Bayesian Optimization tuner
tuner = kt.BayesianOptimization(build_model,
                                objective='val_mae',
                                max_trials=50,
                                directory='bayesian_optimization',
                                project_name='time2vec_rnn')

stop_early = EarlyStopping(monitor='val_loss', patience=5)

# Performing the hyperparameter search
tuner.search(X_train_np, Y_train_np, epochs=50, validation_data=(X_test_np, Y_test_np), callbacks=[stop_early])

# Retrieving the best model
best_model = tuner.get_best_models(num_models=1)[0]

# Verifying the best hyperparameters
tuner.get_best_hyperparameters()[0].values

# Storing the best model
best_model.save('/Users/manotas/Documents/GitHub-Repos/ML-Energy-Colombia/models/non_overlapping_NN')

### 2. Predictions and metrics

# Loading the best model
best_model = load_model('/Users/manotas/Documents/GitHub-Repos/ML-Energy-Colombia/models/non_overlapping_NN')

#### Predictions on train dataset

y_pred_train = np.squeeze(best_model.predict(X_train_np), axis=-1)

#### On test dataset

y_pred_test = np.squeeze(best_model.predict(X_test_np), axis=-1)

#### Comparison

print('Best non-overlapping RNN model - Train metrics:')
calculate_metrics(Y_train_np, y_pred_train)
print('\n Best non-overlapping RNN model - Test metrics:')
calculate_metrics(Y_test_np, y_pred_test)
